WEBDOM-391: Implement https://github.com/digipolisgent/symfony_bundle_domainator9k-core/pull/8.
