+++ Tests/Provisioner/AbstractProvisionerTest.php
@@ -1,15 +1,15 @@
 <?php
 
 
-namespace DigipolisGent\Domainator9k\ServerTypes\CapistranoOpenmindsBundle\Tests\EventListener;
+namespace DigipolisGent\Domainator9k\ServerTypes\CapistranoOpenmindsBundle\Tests\Provisioner;
 
 use DigipolisGent\Domainator9k\CoreBundle\Entity\ApplicationEnvironment;
 use DigipolisGent\Domainator9k\CoreBundle\Entity\Environment;
 use DigipolisGent\Domainator9k\CoreBundle\Entity\Task;
 use DigipolisGent\Domainator9k\CoreBundle\Service\TaskService;
 use DigipolisGent\Domainator9k\CoreBundle\Service\TemplateService;
-use DigipolisGent\Domainator9k\ServerTypes\CapistranoOpenmindsBundle\EventListener\AbstractEventListener;
-use DigipolisGent\Domainator9k\ServerTypes\CapistranoOpenmindsBundle\EventListener\BuildEventListener;
+use DigipolisGent\Domainator9k\ServerTypes\CapistranoOpenmindsBundle\Provisioner\AbstractProvisioner;
+use DigipolisGent\Domainator9k\ServerTypes\CapistranoOpenmindsBundle\Provisioner\BuildProvisioner;
 use DigipolisGent\Domainator9k\ServerTypes\CapistranoOpenmindsBundle\Tests\Fixtures\FooApplication;
 use DigipolisGent\SettingBundle\Service\DataValueService;
 use Doctrine\ORM\EntityManagerInterface;
@@ -17,7 +17,7 @@ use Doctrine\ORM\EntityRepository;
 use phpseclib\Net\SSH2;
 use PHPUnit\Framework\TestCase;
 
-abstract class AbstractEventistenerTest extends TestCase
+abstract class AbstractProvisionerTest extends TestCase
 {
 
     protected function getSsh2Mock()
@@ -97,7 +97,7 @@ abstract class AbstractEventistenerTest extends TestCase
     /**
      * Call a protected/private method of an event listener.
      *
-     * @param AbstractEventListener &$object
+     * @param AbstractProvisioner &$object
      *   The event listener.
      * @param string $methodName
      *   Name of the method to call.
@@ -107,7 +107,7 @@ abstract class AbstractEventistenerTest extends TestCase
      * @return mixed
      *   Method return.
      */
-    protected function invokeEventListenerMethod(AbstractEventListener $listener, $methodName)
+    protected function invokeProvisionerMethod(AbstractProvisioner $listener, $methodName)
     {
         $environment = new Environment();
         $environment->setName('test');
@@ -119,7 +119,7 @@ abstract class AbstractEventistenerTest extends TestCase
         $applicationEnvironment->setApplication($application);
 
         $task = new Task();
-        $task->setType($listener instanceof BuildEventListener ? Task::TYPE_BUILD : Task::TYPE_DESTROY);
+        $task->setType($listener instanceof BuildProvisioner ? Task::TYPE_BUILD : Task::TYPE_DESTROY);
         $task->setStatus(Task::STATUS_NEW);
         $task->setApplicationEnvironment($applicationEnvironment);
