+++ Tests/Provisioner/DestroyProvisionerTest.php
@@ -1,25 +1,22 @@
 <?php
 
 
-namespace DigipolisGent\Domainator9k\ServerTypes\CapistranoOpenmindsBundle\Tests\EventListener;
+namespace DigipolisGent\Domainator9k\ServerTypes\CapistranoOpenmindsBundle\Tests\Provisioner;
 
 use DigipolisGent\Domainator9k\CoreBundle\Entity\ApplicationEnvironment;
 use DigipolisGent\Domainator9k\CoreBundle\Entity\Environment;
 use DigipolisGent\Domainator9k\CoreBundle\Entity\Task;
 use DigipolisGent\Domainator9k\CoreBundle\Entity\VirtualServer;
-use DigipolisGent\Domainator9k\CoreBundle\Event\BuildEvent;
-use DigipolisGent\Domainator9k\ServerTypes\CapistranoOpenmindsBundle\Entity\CapistranoFile;
 use DigipolisGent\Domainator9k\ServerTypes\CapistranoOpenmindsBundle\Entity\CapistranoFolder;
 use DigipolisGent\Domainator9k\ServerTypes\CapistranoOpenmindsBundle\Entity\CapistranoSymlink;
-use DigipolisGent\Domainator9k\ServerTypes\CapistranoOpenmindsBundle\EventListener\AbstractEventListener;
-use DigipolisGent\Domainator9k\ServerTypes\CapistranoOpenmindsBundle\EventListener\BuildEventListener;
+use DigipolisGent\Domainator9k\ServerTypes\CapistranoOpenmindsBundle\Provisioner\DestroyProvisioner;
 use DigipolisGent\Domainator9k\ServerTypes\CapistranoOpenmindsBundle\Tests\Fixtures\FooApplication;
 use Doctrine\Common\Collections\ArrayCollection;
 
-class BuildEventistenerTest extends AbstractEventistenerTest
+class DestroyProvisionerTest extends AbstractProvisionerTest
 {
 
-    public function testOnBuild()
+    public function testOnDestroy()
     {
         $application = new FooApplication();
 
@@ -42,13 +39,11 @@ class BuildEventistenerTest extends AbstractEventistenerTest
         $servers->add($server);
 
         $task = new Task();
-        $task->setType(Task::TYPE_BUILD);
+        $task->setType(Task::TYPE_DESTROY);
         $task->setStatus(Task::STATUS_NEW);
         $task->setApplicationEnvironment($applicationEnvironment);
 
-        $event = new BuildEvent($task);
-
-        $dataValueService = $this->getDataValueServiceMock([true,'user']);
+        $dataValueService = $this->getDataValueServiceMock([true,'username']);
         $templateService = $this->getTemplateServiceMock();
         $taskService = $this->getTaskServiceMock();
         $entityManager = $this->getEntityManagerMock();
@@ -71,10 +66,13 @@ class BuildEventistenerTest extends AbstractEventistenerTest
             'getSshCommand' => function () {
                 return $this->getSsh2Mock();
             },
-            'createFolders' => function () {
+            'removeFiles' => function () {
                 return null;
             },
-            'createSymlinks' => function () {
+            'removeSymlinks' => function () {
+                return null;
+            },
+            'removeFolders' => function () {
                 return null;
             },
             'createFiles' => function () {
@@ -82,11 +80,52 @@ class BuildEventistenerTest extends AbstractEventistenerTest
             }
         ];
 
-        $eventListener = $this->getEventListenerMock($arguments, $methods);
-        $eventListener->onBuild($event);
+        $provisioner = $this->getProvisionerMock($arguments, $methods);
+        $provisioner->run($task);
     }
 
-    public function testCreateFolders()
+
+    public function testRemoveSymlinks()
+    {
+        $dataValueService = $this->getDataValueServiceMock([]);
+        $templateService = $this->getTemplateServiceMock();
+        $taskService = $this->getTaskServiceMock();
+        $entityManager = $this->getEntityManagerMock();
+
+        $symlinks = new ArrayCollection();
+        $symlink = new CapistranoSymlink();
+        $symlink->setSourceLocation('/path/to/my/source');
+        $symlinks->add($symlink);
+
+        $dataValueService
+            ->expects($this->at(0))
+            ->method('getValue')
+            ->willReturn($symlinks);
+
+        $templateService
+            ->expects($this->at(0))
+            ->method('replaceKeys')
+            ->willReturn('/path/to/my/source');
+
+        $provisioner = new DestroyProvisioner(
+            $dataValueService,
+            $templateService,
+            $taskService,
+            $entityManager
+        );
+
+        $source = escapeshellarg('/path/to/my/source');
+        $ssh = $this->getSsh2Mock();
+        $ssh->expects($this->at(0))
+            ->method('exec')
+            ->with($this->equalTo('rm ' . $source));
+
+        $applicationEnvironment = new ApplicationEnvironment();
+
+        $this->invokeProvisionerMethod($provisioner, 'removeSymlinks', $ssh, $applicationEnvironment);
+    }
+
+    public function testRemoveFolders()
     {
         $dataValueService = $this->getDataValueServiceMock([]);
         $templateService = $this->getTemplateServiceMock();
@@ -108,7 +147,7 @@ class BuildEventistenerTest extends AbstractEventistenerTest
             ->method('replaceKeys')
             ->willReturn('/path/to/my/location');
 
-        $eventListener = new BuildEventListener(
+        $provisioner = new DestroyProvisioner(
             $dataValueService,
             $templateService,
             $taskService,
@@ -119,119 +158,18 @@ class BuildEventistenerTest extends AbstractEventistenerTest
         $ssh = $this->getSsh2Mock();
         $ssh->expects($this->at(0))
             ->method('exec')
-            ->with($this->equalTo('mkdir -p -m 750 ' . $path));
+            ->with($this->equalTo('rm -rf ' . $path));
 
 
         $applicationEnvironment = new ApplicationEnvironment();
 
-        $this->invokeEventListenerMethod($eventListener, 'createFolders', $ssh, $applicationEnvironment);
+        $this->invokeProvisionerMethod($provisioner, 'removeFolders', $ssh, $applicationEnvironment);
     }
 
-    public function testCreateSymlinks()
-    {
-        $dataValueService = $this->getDataValueServiceMock([]);
-        $templateService = $this->getTemplateServiceMock();
-        $taskService = $this->getTaskServiceMock();
-        $entityManager = $this->getEntityManagerMock();
-
-        $symlinks = new ArrayCollection();
-        $symlink = new CapistranoSymlink();
-        $symlink->setDestinationLocation('/path/to/my/destination');
-        $symlink->setSourceLocation('/path/to/my/source');
-        $symlinks->add($symlink);
-
-        $dataValueService
-            ->expects($this->at(0))
-            ->method('getValue')
-            ->willReturn($symlinks);
-
-        $templateService
-            ->expects($this->at(0))
-            ->method('replaceKeys')
-            ->willReturn('/path/to/my/source');
-
-        $templateService
-            ->expects($this->at(1))
-            ->method('replaceKeys')
-            ->willReturn('/path/to/my/destination');
-
-        $eventListener = new BuildEventListener(
-            $dataValueService,
-            $templateService,
-            $taskService,
-            $entityManager
-        );
-
-        $destination = escapeshellarg('/path/to/my/destination');
-        $source = escapeshellarg('/path/to/my/source');
-
-        $ssh = $this->getSsh2Mock();
-        $ssh->expects($this->at(0))
-            ->method('exec')
-            ->with($this->equalTo('ln -sfn ' . $destination . ' ' . $source));
-
-        $applicationEnvironment = new ApplicationEnvironment();
-
-        $this->invokeEventListenerMethod($eventListener, 'createSymlinks', $ssh, $applicationEnvironment);
-    }
-
-    public function testCreateFiles()
-    {
-        $dataValueService = $this->getDataValueServiceMock([]);
-        $templateService = $this->getTemplateServiceMock();
-        $taskService = $this->getTaskServiceMock();
-        $entityManager = $this->getEntityManagerMock();
-
-        $files = new ArrayCollection();
-        $file = new CapistranoFile();
-        $file->setLocation('/path/to/location');
-        $file->setFilename('file');
-        $file->setExtension('ext');
-        $file->setContent('my content');
-        $files->add($file);
-
-        $dataValueService
-            ->expects($this->at(0))
-            ->method('getValue')
-            ->willReturn($files);
-
-        $templateService
-            ->expects($this->at(0))
-            ->method('replaceKeys')
-            ->willReturn('/path/to/location/file.ext');
-
-        $templateService
-            ->expects($this->at(1))
-            ->method('replaceKeys')
-            ->willReturn('my-content');
-
-        $eventListener = new BuildEventListener(
-            $dataValueService,
-            $templateService,
-            $taskService,
-            $entityManager
-        );
-
-        $path = escapeshellarg('/path/to/location/file.ext');
-        $content = escapeshellarg('my-content');
-        $ssh = $this->getSsh2Mock();
-        $ssh->expects($this->at(0))
-            ->method('exec')
-            ->with($this->equalTo(
-                "[[ ! -f $path ]] || MOD=$(stat --format '%a' $path "
-              . "&& chmod 600 $path) "
-              . "&& echo $content > $path "
-              . "&& [[ -z \"\$MOD\" ]] || chmod \$MOD $path"));
-
-        $applicationEnvironment = new ApplicationEnvironment();
-
-        $this->invokeEventListenerMethod($eventListener, 'createFiles', $ssh, $applicationEnvironment);
-    }
-
-    private function getEventListenerMock(array $arguments, array $methods)
+    private function getProvisionerMock(array $arguments, array $methods)
     {
         $mock = $this
-            ->getMockBuilder(BuildEventListener::class)
+            ->getMockBuilder(DestroyProvisioner::class)
             ->setMethods(array_keys($methods))
             ->setConstructorArgs($arguments)
             ->getMock();
